rag_backend/requirements.txt
fastapi
uvicorn
faiss-cpu
requests
numpy
pydantic
pdfplumber
langchain
